+++ src/controllers/scrapeJob.js
@@ -43,6 +43,8 @@ function ScrapeJobController(context) {
 
   const HEADER_ERROR = 'x-error';
 
+  const MAX_JOBS_BY_BASEURL = 100;
+
   function createErrorResponse(error) {
     return createResponse({}, error.status || 500, {
       [HEADER_ERROR]: error.message,
@@ -79,6 +81,8 @@ function ScrapeJobController(context) {
       endDate: requestContext.params.endDate,
       baseURL: requestContext.params.baseURL,
       processingType: requestContext.params.processingType,
+      url: requestContext.params.url,
+      maxAge: requestContext.params.maxAge,
     };
   }
 
@@ -182,9 +186,39 @@ function ScrapeJobController(context) {
       }
 
       // return the latest max 100 jobs, sorted by startedAt (newest first)
-      return ok(jobs.sort((a, b) => new Date(b.startedAt) - new Date(a.startedAt)).slice(0, 100));
+      return ok(jobs.sort(
+        (a, b) => new Date(b.startedAt) - new Date(a.startedAt),
+      ).slice(0, MAX_JOBS_BY_BASEURL));
     } catch (error) {
       log.error(`Failed to fetch scrape jobs by baseURL: ${decodedBaseURL}, ${error.message}`);
+      return createErrorResponse(error);
+    }
+  }
+
+  async function getScrapeUrlByProcessingType(requestContext) {
+    const { url: encodedUrl, processingType } = parseRequestContext(requestContext);
+
+    if (!hasText(encodedUrl)) {
+      return badRequest('A valid URL is required');
+    } else if (!hasText(processingType)) {
+      return badRequest('A processing type is required');
+    }
+
+    let decodedUrl = encodedUrl;
+    try {
+      decodedUrl = Buffer.from(encodedUrl, 'base64').toString('utf-8').trim();
+      const scrapeUrls = await scrapeClient.getScrapeUrlsByProcessingType(
+        decodedUrl,
+        processingType,
+      );
+
+      if (!scrapeUrls || scrapeUrls.length === 0) {
+        return ok([]);
+      }
+
+      return ok(scrapeUrls.sort((a, b) => new Date(b.createdAt) - new Date(a.createdAt)));
+    } catch (error) {
+      log.error(`Failed to fetch scrape URLs for url: ${decodedUrl} and processingType: ${processingType}, ${error.message}`);
       return createErrorResponse(error);
     }
   }
@@ -195,6 +229,7 @@ function ScrapeJobController(context) {
     getScrapeJobUrlResults,
     getScrapeJobsByBaseURL,
     getScrapeJobsByDateRange,
+    getScrapeUrlByProcessingType,
   };
 }